自动生成bean，将接口返回值放入bean中，当返回值字段名称&类型都变动时，只修改bean，不对用例操作

+++ src/main/java/com/haisheng/framework/testng/bigScreen/jiaochen/wm/sense/pc/integralcenter/EditExchangeGoodsScene.java
@@ -11,13 +11,13 @@ import lombok.Builder;
 public class EditExchangeGoodsScene extends BaseScene {
     private final String exchangeGoodsType;
     private final Long goodsId;
-    private final Integer exchangePrice;
+    private final Long exchangePrice;
     private final Boolean isLimit;
     private final Integer exchangePeopleNum;
     private final String exchangeStartTime;
     private final String exchangeEndTime;
     private final Long id;
-    private final Integer exchangeNum;
+    private final Long exchangeNum;
     private final Integer expireType;
     private final Integer useDays;
